requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Deprecated==1.2.18
Django==5.2.1
django-background-tasks==1.2.8
django-prometheus==2.3.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
opentelemetry-api==1.33.1
opentelemetry-distro==0.54b1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-dbapi==0.54b1
opentelemetry-instrumentation-django==0.54b1
opentelemetry-instrumentation-sqlite3==0.54b1
opentelemetry-instrumentation-wsgi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
packaging==25.0
prometheus_client==0.22.0
protobuf==5.29.4
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
zipp==3.21.0
